cleanup and account for multiple levers being changed
